% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{acc}
\alias{acc}
\title{Getting the accuracy}
\usage{
acc(y_true, y_hat)
}
\arguments{
\item{y_true}{Integer vector of true binary labels (0/1).}
\item{y_hat}{Integer vector of predicted binary labels (0/1).}
}
\description{
Getting the accuracy
}
