% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{cr_coverage}
\alias{cr_coverage}
\title{A function to calculate coverage of the credible interval}
\usage{
cr_coverage(f_true, f_post, prob = 0.5)
}
\arguments{
\item{f_true}{Numeric vector of true function values.}
\item{f_post}{Matrix of posterior samples (n x n_mcmc).}
\item{prob}{Nominal coverage probability (default 0.5).}
}
\description{
A function to calculate coverage of the credible interval
}
