% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{crps}
\alias{crps}
\title{Calculating CRPS from (https://arxiv.org/pdf/1709.04743.pdf)}
\usage{
crps(y, means, sds)
}
\arguments{
\item{y}{Numeric vector of observed values.}
\item{means}{Numeric vector of predictive means.}
\item{sds}{Numeric vector of predictive standard deviations.}
}
\description{
Calculating CRPS from (https://arxiv.org/pdf/1709.04743.pdf)
}
