% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{logloss}
\alias{logloss}
\title{Binary classification metrics}
\usage{
logloss(y_true, y_hat)
}
\arguments{
\item{y_true}{Integer vector of true binary labels (0/1).}
\item{y_hat}{Numeric vector of predicted probabilities.}
}
\description{
Binary classification metrics
}
