% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{recode_vars}
\alias{recode_vars}
\title{Recoding variables}
\usage{
recode_vars(x_train, dummy_obj)
}
\arguments{
\item{x_train}{A \code{data.frame} of training covariates after dummy-variable expansion.}
\item{dummy_obj}{A dummy-variable object returned by \code{caret::dummyVars}.}
}
\description{
Recoding variables
}
