% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{rmse}
\alias{rmse}
\title{RMSE: Calculating the rmse}
\usage{
rmse(x, y)
}
\arguments{
\item{x}{Numeric vector of predicted values.}
\item{y}{Numeric vector of observed values.}
}
\description{
RMSE: Calculating the rmse
}
