% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subart.R
\name{subart}
\alias{subart}
\title{Multivariate Normal Bayesian Additive Regression trees.}
\usage{
subart(
  x_train,
  y_mat,
  x_test = NULL,
  n_tree = 100,
  node_min_size = 5,
  n_mcmc = 2000,
  n_burn = 500,
  alpha = 0.95,
  beta = 2,
  nu = 3,
  sigquant = 0.9,
  kappa = 2,
  numcut = 100L,
  usequants = FALSE,
  m = 20,
  hier_prior_sigma = TRUE,
  diagnostic = TRUE
)
}
\arguments{
\item{x_train}{A \code{data.frame} of the training data covariates.}

\item{y_mat}{A numeric matrix of the data responses.}

\item{x_test}{A \code{data.frame} of the test data covariates.}

\item{n_tree}{The number of trees used in each set of trees for the respective \eqn{j} entry. The total number of trees is given by \eqn{j \times p}.}

\item{node_min_size}{The minimun number of observations within a terminal node.}

\item{n_mcmc}{The total number of MCMC iterations.}

\item{n_burn}{The number of MCMC iterations to be tretead as burn in samples}

\item{alpha}{The power parameter used in tree prior.}

\item{beta}{The base parameter used in tree prior.}

\item{nu}{The \eqn{\nu} parameter associated with the degree of freedom from the variance prior.}

\item{sigquant}{Quantile used to define the residuals variance. Remind that the prior definition is based on \eqn{P(\sigma^{(j)} < \hat{\sigma}^{(j)})} where \eqn{\hat{\sigma}^{(j)}} is a "rough data-based estimation" for example, the sample variance of the observed \eqn{y^{(j)}} values.}

\item{kappa}{Hyper-parameter from the \eqn{\mu_{t\ell}^{(j)}} (Chipman et. al 2010). Such that \eqn{\sigma} = 0.25}

\item{numcut}{The maximum number of possible values used in the tree decision rules. The uniform approximation for choose a decision rule over \eqn{X^{(j)}} is given a grid of size \code{numcut}.}

\item{usequants}{Boolean; if true the quantiles are going to be used to define the grid of cutpoints.}

\item{m}{Hyperparameter used in the definition of the prior setting of the correlation matrix for the Probit-Multivariate approach.}

\item{hier_prior_sigma}{Boolean; if true the prior for Sigma is defined using the hierachical prior as defined by Esser & Maia et. al 2025. See details for reference.}

\item{diagnostic}{a boolean to compute or not the ESS for the posterior samples of the \eqn{\boldsymbol{\Sigma}}}
}
\value{
In case of a continuous response the model returns
\code{subart} object with the model predictions and parameters for the standard MVN. For the binary-classification problems the probit
Probit-MVN approach is used, and returns the \code{subart} object.
}
\description{
\code{subart()} function models a Bayesian Additive Regression trees model considering the Multivariate Normal (MVN) distribution
from the target dependent variable \eqn{Y_{i} \in \mathbb{R}^{d}}.
}
\details{
Add more details of the function description
}
